+++ packages/react/src/components/button/button-group.tsx
@@ -7,8 +7,8 @@ import type { ButtonProps } from "./button"
 import type { ButtonStyle } from "./button.style"
 import { useMemo } from "react"
 import { Group } from "../group"
+import { IconButtonPropsContext } from "../icon-button"
 import { ButtonPropsContext } from "./button"
-import { IconButtonPropsContext } from "./icon-button"
 
 export interface ButtonGroupProps
   extends WithoutThemeProps<GroupProps, ButtonStyle>,
