+++ packages/react/src/components/editable/editable.stories.tsx
@@ -4,9 +4,9 @@ import { Controller, useForm } from "react-hook-form"
 import { Editable } from "."
 import { Button } from "../button"
 import { ButtonGroup } from "../button"
-import { IconButton } from "../button"
 import { Field } from "../field"
 import { CheckIcon, PencilIcon, XIcon } from "../icon"
+import { IconButton } from "../icon-button"
 import { VStack } from "../stack"
 
 type Story = StoryFn<typeof Editable.Root>
