+++ packages/react/src/components/file-button/file-button.stories.tsx
@@ -8,9 +8,9 @@ import { PlusIcon } from "../../components/icon"
 import { HStack, VStack } from "../../components/stack"
 import { Text } from "../../components/text"
 import { COLOR_SCHEMES, toTitleCase } from "../../utils"
-import { IconButton } from "../button"
 import { Field } from "../field"
 import { For } from "../for"
+import { IconButton } from "../icon-button"
 import { Wrap } from "../wrap"
 import { FileButton } from "./file-button"
