+++ packages/react/src/components/group/group.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryFn } from "@storybook/react-vite"
 import { Button } from "../button"
-import { IconButton } from "../button"
 import { PlusIcon } from "../icon"
+import { IconButton } from "../icon-button"
 import { Input } from "../input"
 import { Tag } from "../tag"
 import { Group } from "./group"
