+++ packages/react/src/components/icon-button/index.ts
@@ -1,0 +1,7 @@
+export {
+  IconButton,
+  IconButtonPropsContext,
+  useIconButtonPropsContext,
+} from "./icon-button"
+export type { IconButtonProps } from "./icon-button"
+export * from "./icon-button.style"
