+++ packages/react/src/components/number-input/number-input.stories.tsx
@@ -4,10 +4,10 @@ import { Controller, useForm } from "react-hook-form"
 import { PropsTable } from "../../../storybook/components"
 import { COLOR_SCHEMES, toTitleCase } from "../../utils"
 import { Button } from "../button"
-import { IconButton } from "../button"
 import { Field } from "../field"
 import { For } from "../for"
 import { GaugeIcon, MinusIcon, PlusIcon } from "../icon"
+import { IconButton } from "../icon-button"
 import { Input, InputGroup } from "../input"
 import { HStack, VStack } from "../stack"
 import { NumberInput, useNumberInput } from "./"
