+++ packages/react/src/components/pagination/pagination.stories.tsx
@@ -3,7 +3,6 @@ import { useState } from "react"
 import { Pagination } from "."
 import { PropsTable } from "../../../storybook/components"
 import { COLOR_SCHEMES, isNumber } from "../../utils"
-import { IconButton } from "../button"
 import {
   ArrowLeftIcon,
   ArrowLeftToLineIcon,
@@ -17,6 +16,7 @@ import {
   GhostIcon,
   MinusIcon,
 } from "../icon"
+import { IconButton } from "../icon-button"
 
 type Story = StoryFn<typeof Pagination.Root>
