+++ packages/react/src/components/pagination/pagination.tsx
@@ -9,7 +9,8 @@ import type {
   WithoutThemeProps,
 } from "../../core"
 import type { ReactNodeOrFunction } from "../../utils"
-import type { ButtonGroupProps, IconButtonProps } from "../button"
+import type { ButtonGroupProps } from "../button"
+import type { IconButtonProps } from "../icon-button"
 import type { PaginationStyle } from "./pagination.style"
 import type { Page, UsePaginationProps } from "./use-pagination"
 import { cloneElement, isValidElement, useMemo } from "react"
@@ -17,7 +18,7 @@ import { createSlotComponent, styled } from "../../core"
 import { useValue } from "../../hooks/use-value"
 import { useI18n } from "../../providers/i18n-provider"
 import { isNumber, runIfFn } from "../../utils"
-import { ButtonGroup, IconButton } from "../button"
+import { ButtonGroup } from "../button"
 import {
   ChevronLeftIcon,
   ChevronRightIcon,
@@ -25,6 +26,7 @@ import {
   ChevronsRightIcon,
   EllipsisIcon,
 } from "../icon"
+import { IconButton } from "../icon-button"
 import { paginationStyle } from "./pagination.style"
 import {
   PaginationContext,
