+++ packages/react/src/components/toggle/toggle.style.ts
@@ -1,5 +1,5 @@
 import { defineComponentSlotStyle } from "../../core"
-import { iconButtonStyle } from "../button"
+import { iconButtonStyle } from "../icon-button"
 
 export const toggleStyle = defineComponentSlotStyle({
   base: {
