+++ packages/react/src/index.ts
@@ -48,6 +48,7 @@ export * from "./components/group"
 export * from "./components/heading"
 export * from "./components/highlight"
 export * from "./components/icon"
+export * from "./components/icon-button"
 export * from "./components/image"
 export * from "./components/indicator"
 export * from "./components/infinite-scroll-area"
